completed navbar and corrected file names with snake casing, fix routing and css styling for nav bar.
