+++ client/src/components/bulletin_board.css
