+++ client/src/components/forkNav.js
@@ -1,18 +1,13 @@
 import React from "react";
-import {Link} from 'react-router-dom';
-import blueLeftArrow from "./images/blue-chevron-left.png";
+import { Link } from "react-router-dom";
+import backArrow from "./images/blue-chevron-left.png";
+import Navbar from "./navbar";
+import hamburgerMenu from "./hamburger_menu";
 import "./styles.css";
 
 export default props => (
 	<div className="forkContainer">
-		<div className="backNav">
-			<div className="backArrow">
-				<img className="backArrowImg" src={blueLeftArrow} />
-			</div>
-			<div className="backTsLogo">
-				<span className="backNavTitle">Team Stream</span>
-			</div>
-		</div>
+		<Navbar icon={backArrow} hamburgerMenu={false} />
 		<div className="forkButtonContent">
 			<Link to="/create_team" className="forkButtons">
 				<span className="btnFork">Create Team</span>
