+++ client/src/components/pin_message.js
